compiler warning fixes + one real bug in dcc-chat
git-svn-id: http://svn.irssi.org/repos/irssi/trunk@850 dbcabf3a-b0e7-0310-adc4-f8d773084564

+++ src/core/levels.c
@@ -66,7 +66,7 @@ int level_get(const char *level)
 	match = 0;
 	for (n = 0; levels[n] != NULL; n++) {
 		if (strncmp(levels[n], level, len) == 0) {
-			if (strlen(levels[n]) == len) {
+			if ((int)strlen(levels[n]) == len) {
 				/* full match */
 				return 1L << n;
 			}
